move folders inside src
