+++ server/src/controllers/auth.js
@@ -1,7 +1,7 @@
 import User from "../models/User.js";
 
 const sendTokenResponse = (user, statusCode, res) => {
-//create token
+  //create token
   const token = user.getSignedJwtToken();
 
   const options = {
@@ -18,7 +18,7 @@ const sendTokenResponse = (user, statusCode, res) => {
   res
     .status(statusCode)
     // .cookie("token", token, options)
-    .json({ success: true, _id:user._id, name: user.nickname, token });
+    .json({ success: true, _id: user._id, name: user.nickname, token });
 };
 
 // @desc    Register a user
@@ -32,13 +32,15 @@ export const register = async (req, res, next) => {
     const user = await User.create({
       nickname,
       username,
-      password
+      password,
     });
 
     sendTokenResponse(user, 200, res);
   } catch (err) {
-    if(err.code === 11000) {
-      return res.status(400).json({ success: false, msg: "Username already exists" });
+    if (err.code === 11000) {
+      return res
+        .status(400)
+        .json({ success: false, msg: "Username already exists" });
     }
     res.status(400).json({ success: false });
     console.log(err.stack);
@@ -50,35 +52,39 @@ export const register = async (req, res, next) => {
 // @access  Public
 export const login = async (req, res, next) => {
   try {
-  const { username, password } = req.body;
+    const { username, password } = req.body;
 
-//validate
-  if (!username || !password) {
-    return res
-      .status(400)
-      .json({ success: false, msg: "Please provide an username and password" });
-  }
-//check for user
-  const user = await User.findOne({ username }).select("+password");
+    //validate
+    if (!username || !password) {
+      return res
+        .status(400)
+        .json({
+          success: false,
+          msg: "Please provide an username and password",
+        });
+    }
+    //check for user
+    const user = await User.findOne({ username }).select("+password");
 
-  if (!user) {
-    return res
-      .status(401)
-      .json({ success: false, msg: "False Username" });
-  }
+    if (!user) {
+      return res.status(401).json({ success: false, msg: "False Username" });
+    }
 
-//check if password match
-  const isMatch = await user.matchPassword(password);
+    //check if password match
+    const isMatch = await user.matchPassword(password);
 
-  if (!isMatch) {
-    return res
-      .status(401)
-      .json({ success: false, msg: "False Password" });
-  }
+    if (!isMatch) {
+      return res.status(401).json({ success: false, msg: "False Password" });
+    }
 
     sendTokenResponse(user, 200, res);
   } catch (err) {
-    res.status(401).json({ success: false, msg: "Cannot convert email or password to string" });
+    res
+      .status(401)
+      .json({
+        success: false,
+        msg: "Cannot convert email or password to string",
+      });
   }
 };
 
@@ -87,13 +93,13 @@ export const login = async (req, res, next) => {
 // @access  Private
 export const getMe = async (req, res, next) => {
   const user = await User.findById(req.user.id);
-    res.status(200).json({
-        success: true,
-        data: user
-    });
-}
+  res.status(200).json({
+    success: true,
+    data: user,
+  });
+};
 
-export const logout  = async (req, res, next) => {
+export const logout = async (req, res, next) => {
   res.cookie("token", "none", {
     expires: new Date(Date.now() + 10 * 1000),
     httpOnly: true,
@@ -102,4 +108,4 @@ export const logout  = async (req, res, next) => {
     success: true,
     data: {},
   });
-}
+};
