+++ server/src/middleware/auth.js
@@ -1,0 +1,46 @@
+import jwt from "jsonwebtoken";
+import User from "../models/User.js";
+
+export const protect = async (req, res, next) => {
+  let token;
+
+  if (
+    req.headers.authorization &&
+    req.headers.authorization.startsWith("Bearer")
+  ) {
+    token = req.headers.authorization.split(" ")[1];
+  }
+
+  if (!token || token === "null") {
+    return res
+      .status(401)
+      .json({ success: false, error: "Not authorized to access this route" });
+  }
+
+  try {
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+
+    console.log(decoded);
+
+    req.user = await User.findById(decoded.id);
+
+    next();
+  } catch (err) {
+    console.log(err.stack);
+    return res
+      .status(401)
+      .json({ success: false, error: "Not authorized to access this route" });
+  }
+};
+
+export const authorize = (...roles) => {
+  return (req, res, next) => {
+    if (!roles.includes(req.user.role)) {
+      return res.status(403).json({
+        success: false,
+        error: `User role ${req.user.role} is not authorized to access this route`,
+      });
+    }
+    next();
+  };
+};
