+++ server/src/server.js
@@ -3,7 +3,8 @@ import express from "express";
 import { Server } from "socket.io";
 import dotenv from "dotenv";
 import cors from "cors";
-import connectDB from "../database/db.js";
+import connectDB from "./database/db.js";
+import auth from "./routes/auth.js";
 
 dotenv.config();
 connectDB();
@@ -20,8 +21,6 @@ app.use(
 );
 
 app.use(express.json());
-import auth from "../routes/auth.js";
-
 app.use("/api/v1/auth", auth);
 
 const server = http.createServer(app);
